util: move device filtering
With multiple functions filtering out MAJORs of 1 or 7, move this
function to utils and use it.
